+++ src/test/java/uk/gov/pay/api/it/GetPaymentITest.java
@@ -27,7 +27,7 @@ import static org.hamcrest.core.Is.is;
 import static uk.gov.pay.api.model.TokenPaymentType.CARD;
 import static uk.gov.pay.api.model.TokenPaymentType.DIRECT_DEBIT;
 
-public class PaymentsResourceITest extends PaymentResourceITestBase {
+public class GetPaymentITest extends PaymentResourceITestBase {
 
     private static final ZonedDateTime TIMESTAMP = DateTimeUtils.toUTCZonedDateTime("2016-01-01T12:00:00Z").get();
     private static final ZonedDateTime CAPTURED_DATE = ZonedDateTime.parse("2016-01-02T14:03:00Z");
@@ -50,159 +50,6 @@ public class PaymentsResourceITest extends PaymentResourceITestBase {
     private static final List<Map<String, String>> EVENTS = Collections.singletonList(PAYMENT_CREATED);
     private static final Address BILLING_ADDRESS = new Address("line1", "line2", "NR2 5 6EG", "city", "UK");
     private static final CardDetails CARD_DETAILS = new CardDetails("1234", "Mr. Payment", "12/19", BILLING_ADDRESS, CARD_BRAND_LABEL);
-    private static final String SUCCESS_PAYLOAD = paymentPayload(AMOUNT, RETURN_URL, DESCRIPTION, REFERENCE, EMAIL);
-
-    @Test
-    public void createCardPayment() {
-
-        publicAuthMock.mapBearerTokenToAccountId(API_KEY, GATEWAY_ACCOUNT_ID, CARD);
-
-        connectorMock.respondOk_whenCreateCharge(AMOUNT, GATEWAY_ACCOUNT_ID, CHARGE_ID, CHARGE_TOKEN_ID,
-                CREATED, RETURN_URL, DESCRIPTION, REFERENCE, null, PAYMENT_PROVIDER, CREATED_DATE, REFUND_SUMMARY, null, CARD_DETAILS);
-
-        String responseBody = postPaymentResponse(API_KEY, SUCCESS_PAYLOAD)
-                .statusCode(201)
-                .contentType(JSON)
-                .header(HttpHeaders.LOCATION, is(paymentLocationFor(CHARGE_ID)))
-                .body("payment_id", is(CHARGE_ID))
-                .body("amount", is(9999999))
-                .body("reference", is(REFERENCE))
-                .body("email", nullValue())
-                .body("description", is(DESCRIPTION))
-                .body("state.status", is(CREATED.getStatus()))
-                .body("return_url", is(RETURN_URL))
-                .body("payment_provider", is(PAYMENT_PROVIDER))
-                .body("card_brand", is(CARD_BRAND_LABEL))
-                .body("created_date", is(CREATED_DATE))
-                .body("refund_summary.status", is("pending"))
-                .body("refund_summary.amount_submitted", is(50))
-                .body("refund_summary.amount_available", is(100))
-                .body("_links.self.href", is(paymentLocationFor(CHARGE_ID)))
-                .body("_links.self.method", is("GET"))
-                .body("_links.next_url.href", is(frontendUrlFor(CARD) + CHARGE_TOKEN_ID))
-                .body("_links.next_url.method", is("GET"))
-                .body("_links.next_url_post.href", is(frontendUrlFor(CARD)))
-                .body("_links.next_url_post.method", is("POST"))
-                .body("_links.next_url_post.type", is("application/x-www-form-urlencoded"))
-                .body("_links.next_url_post.params.chargeTokenId", is(CHARGE_TOKEN_ID))
-                .body("_links.events.href", is(paymentEventsLocationFor(CHARGE_ID)))
-                .body("_links.events.method", is("GET"))
-                .body("_links.cancel.href", is(paymentCancelLocationFor(CHARGE_ID)))
-                .body("_links.cancel.method", is("POST"))
-                .body("_links.refunds.href", is(paymentRefundsLocationFor(CHARGE_ID)))
-                .body("_links.refunds.method", is("GET"))
-                .extract().body().asString();
-
-        JsonAssert.with(responseBody)
-                .assertNotDefined("_links.self.type")
-                .assertNotDefined("_links.self.params")
-                .assertNotDefined("_links.next_url.type")
-                .assertNotDefined("_links.next_url.params")
-                .assertNotDefined("_links.events.type")
-                .assertNotDefined("_links.events.params");
-
-        connectorMock.verifyCreateChargeConnectorRequest(AMOUNT, GATEWAY_ACCOUNT_ID, RETURN_URL, DESCRIPTION, REFERENCE);
-    }
-
-    @Test
-    public void createPayment_withMinimumAmount() {
-
-        int minimumAmount = 1;
-
-        publicAuthMock.mapBearerTokenToAccountId(API_KEY, GATEWAY_ACCOUNT_ID);
-        connectorMock.respondOk_whenCreateCharge(minimumAmount, GATEWAY_ACCOUNT_ID, CHARGE_ID, CHARGE_TOKEN_ID,
-                CREATED, RETURN_URL, DESCRIPTION, REFERENCE, EMAIL, PAYMENT_PROVIDER, CREATED_DATE, REFUND_SUMMARY, null, CARD_DETAILS);
-
-        postPaymentResponse(API_KEY, paymentPayload(minimumAmount, RETURN_URL, DESCRIPTION, REFERENCE, EMAIL))
-                .statusCode(201)
-                .contentType(JSON)
-                .body("payment_id", is(CHARGE_ID))
-                .body("amount", is(minimumAmount))
-                .body("reference", is(REFERENCE))
-                .body("email", is(EMAIL))
-                .body("description", is(DESCRIPTION))
-                .body("return_url", is(RETURN_URL))
-                .body("payment_provider", is(PAYMENT_PROVIDER))
-                .body("created_date", is(CREATED_DATE));
-
-        connectorMock.verifyCreateChargeConnectorRequest(minimumAmount, GATEWAY_ACCOUNT_ID, RETURN_URL, DESCRIPTION, REFERENCE);
-    }
-
-    @Test
-    public void createPayment_withAllFieldsUpToMaxLengthBoundaries_shouldBeAccepted() {
-
-        int amount = 10000000;
-        String reference = randomAlphanumeric(255);
-        String description = randomAlphanumeric(255);
-        String email = randomAlphanumeric(254) + "@mail.fake";
-        String return_url = "https://govdemopay.gov.uk?data=" + randomAlphanumeric(1969);
-
-        publicAuthMock.mapBearerTokenToAccountId(API_KEY, GATEWAY_ACCOUNT_ID);
-        connectorMock.respondOk_whenCreateCharge(amount, GATEWAY_ACCOUNT_ID, CHARGE_ID, CHARGE_TOKEN_ID,
-                CREATED, return_url, description, reference, email, PAYMENT_PROVIDER, CREATED_DATE, REFUND_SUMMARY, null, CARD_DETAILS);
-
-        String body = new JsonStringBuilder()
-                .add("amount", amount)
-                .add("reference", reference)
-                .add("email", email)
-                .add("card_brand", CARD_BRAND_LABEL)
-                .add("description", description)
-                .add("return_url", return_url)
-                .build();
-
-        postPaymentResponse(API_KEY, body)
-                .statusCode(201)
-                .contentType(JSON)
-                .body("payment_id", is(CHARGE_ID))
-                .body("amount", is(amount))
-                .body("reference", is(reference))
-                .body("email", is(email))
-                .body("description", is(description))
-                .body("return_url", is(return_url))
-                .body("payment_provider", is(PAYMENT_PROVIDER))
-                .body("card_brand", is(CARD_BRAND_LABEL))
-                .body("created_date", is(CREATED_DATE));
-    }
-
-    @Test
-    public void createPayment_responseWith500_whenConnectorResponseIsAnUnrecognisedError() throws Exception {
-
-        String gatewayAccountId = "1234567";
-        String errorMessage = "something went wrong";
-
-        publicAuthMock.mapBearerTokenToAccountId(API_KEY, gatewayAccountId);
-
-        connectorMock.respondBadRequest_whenCreateCharge(AMOUNT, gatewayAccountId, errorMessage, RETURN_URL, DESCRIPTION, REFERENCE);
-
-        InputStream body = postPaymentResponse(API_KEY, SUCCESS_PAYLOAD)
-                .statusCode(500)
-                .contentType(JSON).extract()
-                .body().asInputStream();
-
-        JsonAssert.with(body)
-                .assertThat("$.*", hasSize(2))
-                .assertThat("$.code", is("P0198"))
-                .assertThat("$.description", is("Downstream system error"));
-
-        connectorMock.verifyCreateChargeConnectorRequest(AMOUNT, gatewayAccountId, RETURN_URL, DESCRIPTION, REFERENCE);
-    }
-
-    @Test
-    public void createPayment_responseWith500_whenTokenForGatewayAccountIsValidButConnectorResponseIsNotFound() {
-
-        String notFoundGatewayAccountId = "9876545";
-        publicAuthMock.mapBearerTokenToAccountId(API_KEY, notFoundGatewayAccountId);
-
-        connectorMock.respondNotFound_whenCreateCharge(AMOUNT, notFoundGatewayAccountId, RETURN_URL, DESCRIPTION, REFERENCE);
-
-        postPaymentResponse(API_KEY, SUCCESS_PAYLOAD)
-                .statusCode(500)
-                .contentType(JSON)
-                .body("code", is("P0199"))
-                .body("description", is("There is an error with this account. Please contact support"));
-
-        connectorMock.verifyCreateChargeConnectorRequest(AMOUNT, notFoundGatewayAccountId, RETURN_URL, DESCRIPTION, REFERENCE);
-    }
 
     @Test
     public void getPayment_ReturnsPayment() {
@@ -421,32 +268,6 @@ public class PaymentsResourceITest extends PaymentResourceITestBase {
                 .assertThat("$.description", is("Downstream system error"));
     }
 
-    @Test
-    public void createPayment_Returns401_WhenUnauthorised() {
-        publicAuthMock.respondUnauthorised();
-
-        postPaymentResponse(API_KEY, SUCCESS_PAYLOAD)
-                .statusCode(401);
-    }
-
-    @Test
-    public void createPayment_Returns_WhenPublicAuthInaccessible() {
-        publicAuthMock.respondWithError();
-
-        postPaymentResponse(API_KEY, SUCCESS_PAYLOAD)
-                .statusCode(503);
-    }
-
-    private static String paymentPayload(long amount, String returnUrl, String description, String reference, String email) {
-        return new JsonStringBuilder()
-                .add("amount", amount)
-                .add("reference", reference)
-                .add("email", email)
-                .add("description", description)
-                .add("return_url", returnUrl)
-                .build();
-    }
-
     private ValidatableResponse getPaymentResponse(String bearerToken, String paymentId) {
         return given().port(app.getLocalPort())
                 .header(AUTHORIZATION, "Bearer " + bearerToken)
@@ -458,16 +279,6 @@ public class PaymentsResourceITest extends PaymentResourceITestBase {
         return given().port(app.getLocalPort())
                 .header(AUTHORIZATION, "Bearer " + bearerToken)
                 .get(String.format("/v1/payments/%s/events", paymentId))
-                .then();
-    }
-
-    private ValidatableResponse postPaymentResponse(String bearerToken, String payload) {
-        return given().port(app.getLocalPort())
-                .body(payload)
-                .accept(JSON)
-                .contentType(JSON)
-                .header(AUTHORIZATION, "Bearer " + bearerToken)
-                .post(PAYMENTS_PATH)
                 .then();
     }
 }
